src/collectors/google_reviews_collector.py
"""
Google Reviews Collector
Collects reviews from Google My Business / Google Maps
"""

import requests
from bs4 import BeautifulSoup
import os
from datetime import datetime


class GoogleReviewsCollector:
    def __init__(self):
        self.api_key = os.getenv('GOOGLE_API_KEY')

    def collect(self, company_name=None, depth='standard'):
        """
        Collect Google reviews

        Args:
            company_name: Company name to search for
            depth: Research depth (quick, standard, deep)

        Returns:
            List of review dictionaries
        """
        reviews = []

        try:
            if not company_name:
                company_name = os.getenv('COMPANY_NAME')

            if not company_name:
                print("Warning: No company name provided for Google Reviews")
                return reviews

            # Note: In production, you would use Google Places API
            # For now, we'll create a structure that can be extended

            # Determine review limits
            review_limits = {
                'quick': 20,
                'standard': 50,
                'deep': 100
            }
            limit = review_limits.get(depth, 50)

            # Placeholder for Google Places API integration
            # You would need to:
            # 1. Search for place_id using company name
            # 2. Fetch reviews using place_id
            # 3. Parse and structure the data

            print(f"Google Reviews collection requires Google Places API setup")
            print(f"Add your Google API key to .env file")

            # Example structure for when API is connected:
            """
            from googleplaces import GooglePlaces, types

            google_places = GooglePlaces(self.api_key)
            query_result = google_places.text_search(query=company_name)

            if query_result.places:
                place = query_result.places[0]
                place.get_details()

                for review in place.reviews[:limit]:
                    reviews.append({
                        'id': review.get('time'),
                        'text': review.get('text', ''),
                        'rating': review.get('rating', 0),
                        'author': review.get('author_name', 'Anonymous'),
                        'date': datetime.fromtimestamp(review.get('time', 0)).isoformat(),
                        'platform': 'google_reviews',
                        'collected_at': datetime.now().isoformat()
                    })
            """

        except Exception as e:
            print(f"Error collecting Google Reviews: {str(e)}")

        return reviews
