requirements.txt
# Company Perception Research App Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Data Collection & APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
tweepy==4.14.0
google-api-python-client==2.108.0
app-store-scraper==0.3.5
google-play-scraper==1.2.4

# AI & NLP
openai==1.3.7
anthropic==0.8.1
transformers==4.36.0
torch==2.1.1
sentencepiece==0.1.99
langchain==0.1.0
langchain-community==0.0.10

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1
